+++ tests/get_set.rs
@@ -1,6 +1,6 @@
 #![cfg(not(feature = "no_object"))]
 
-use rhai::{Engine, EvalAltResult, INT};
+use rhai::{Engine, EvalAltResult, Scope, INT};
 
 #[test]
 fn test_get_set() -> Result<(), Box<EvalAltResult>> {
@@ -70,7 +70,7 @@ fn test_get_set() -> Result<(), Box<EvalAltResult>> {
 }
 
 #[test]
-fn test_get_set_chain() -> Result<(), Box<EvalAltResult>> {
+fn test_get_set_chain_with_write_back() -> Result<(), Box<EvalAltResult>> {
     #[derive(Clone)]
     struct TestChild {
         x: INT,
@@ -119,16 +119,39 @@ fn test_get_set_chain() -> Result<(), Box<EvalAltResult>> {
     engine.register_get_set("x", TestChild::get_x, TestChild::set_x);
     engine.register_get_set("child", TestParent::get_child, TestParent::set_child);
 
-    engine.register_fn("new_tp", TestParent::new);
+    #[cfg(not(feature = "no_index"))]
+    engine.register_indexer_get_set(
+        |parent: &mut TestParent, _: INT| parent.child.clone(),
+        |parent: &mut TestParent, n: INT, mut new_child: TestChild| {
+            new_child.x *= n;
+            parent.child = new_child;
+        },
+    );
 
+    engine.register_fn("new_tp", TestParent::new);
+    engine.register_fn("new_tc", TestChild::new);
+
+    assert_eq!(engine.eval::<INT>("let a = new_tp(); a.child.x")?, 1);
     assert_eq!(
-        engine.eval::<INT>("let a = new_tp(); a.child.x = 500; a.child.x")?,
-        500
+        engine.eval::<INT>("let a = new_tp(); a.child.x = 42; a.child.x")?,
+        42
     );
 
     assert_eq!(
         engine.eval::<String>("let a = new_tp(); type_of(a)")?,
         "TestParent"
+    );
+
+    #[cfg(not(feature = "no_index"))]
+    assert_eq!(
+        engine.eval::<INT>("let a = new_tp(); let c = new_tc(); c.x = 123; a[2] = c; a.child.x")?,
+        246
+    );
+
+    #[cfg(not(feature = "no_index"))]
+    assert_eq!(
+        engine.eval::<INT>("let a = new_tp(); a[2].x = 42; a.child.x")?,
+        84
     );
 
     Ok(())
@@ -163,6 +186,71 @@ fn test_get_set_op_assignment() -> Result<(), Box<EvalAltResult>> {
         engine.eval::<Num>("let a = new_ts(); a.v += 2; a")?,
         Num(42)
     );
+
+    Ok(())
+}
+
+#[test]
+fn test_get_set_chain_without_write_back() -> Result<(), Box<EvalAltResult>> {
+    #[derive(Debug, Clone)]
+    struct Outer {
+        pub inner: Inner,
+    }
+
+    #[derive(Debug, Clone)]
+    struct Inner {
+        pub value: INT,
+    }
+
+    let mut engine = Engine::new();
+    let mut scope = Scope::new();
+
+    scope.push(
+        "outer",
+        Outer {
+            inner: Inner { value: 42 },
+        },
+    );
+
+    engine
+        .register_type::<Inner>()
+        .register_get_set(
+            "value",
+            |t: &mut Inner| t.value,
+            |_: &mut Inner, new: INT| panic!("Inner::value setter called with {}", new),
+        )
+        .register_type::<Outer>()
+        .register_get_set(
+            "inner",
+            |t: &mut Outer| t.inner.clone(),
+            |_: &mut Outer, new: Inner| panic!("Outer::inner setter called with {:?}", new),
+        );
+
+    #[cfg(not(feature = "no_index"))]
+    engine.register_indexer_get_set(
+        |t: &mut Outer, n: INT| Inner {
+            value: t.inner.value * n,
+        },
+        |_: &mut Outer, n: INT, new: Inner| {
+            panic!("Outer::inner index setter called with {} and {:?}", n, new)
+        },
+    );
+
+    assert_eq!(
+        engine.eval_with_scope::<INT>(&mut scope, "outer.inner.value")?,
+        42
+    );
+
+    #[cfg(not(feature = "no_index"))]
+    assert_eq!(
+        engine.eval_with_scope::<INT>(&mut scope, "outer[2].value")?,
+        84
+    );
+
+    engine.run_with_scope(&mut scope, "print(outer.inner.value)")?;
+
+    #[cfg(not(feature = "no_index"))]
+    engine.run_with_scope(&mut scope, "print(outer[0].value)")?;
 
     Ok(())
 }